# Conflicts:
#	index-server/src/main/java/com/cmri/um/he/index/terminal/controller/TerminalOverviewController.java
#	index-server/src/main/java/com/cmri/um/he/index/terminal/mapper/TerminalOverviewMapper.java
#	index-server/src/main/java/com/cmri/um/he/index/terminal/service/TerminalOverviewService.java

+++ index-server/src/main/java/com/cmri/um/he/index/terminal/dao/TerminalOverviewDao.java
@@ -48,7 +48,7 @@ public class TerminalOverviewDao extends BaseDao {
      * @param month
      * @return
      */
-    public int quaryTotal(Integer id, String month) {
+    public double quaryTotal(Integer id, String month) {
         return terminalOverviewMapper.quaryTotal(id,month);
 
     }
